+++ adapters/AppLovin/AppLovinAdapter/GADMediationAdapterAppLovin.m
@@ -21,11 +21,14 @@
 #import "GADMAdapterAppLovinUtils.h"
 #import "GADMRTBAdapterAppLovinInterstitialRenderer.h"
 #import "GADMWaterfallAppLovinAppOpenRenderer.h"
+#import "GADMWaterfallAppLovinBannerRenderer.h"
 #import "GADMWaterfallAppLovinInterstitialRenderer.h"
 
 @implementation GADMediationAdapterAppLovin {
   /// AppLovin app open ad wrapper.
   GADMWaterfallAppLovinAppOpenRenderer *_waterfallAppOpenRenderer;
+
+  GADMWaterfallAppLovinBannerRenderer *_waterfallBannerRenderer;
 
   /// AppLovin interstitial ad wrapper.
   GADMRTBAdapterAppLovinInterstitialRenderer *_rtbInterstitialRenderer;
@@ -190,6 +193,39 @@
                                                      completionHandler:completionHandler];
                                        [strongSelf->_waterfallAppOpenRenderer loadAd];
                                      }];
+}
+
+// Note: Banner ads are supported by AppLovin only for Waterfall and not for Bidding. So, all banner
+// ad load requests are assumed to be for Waterfall.
+- (void)loadBannerForAdConfiguration:(GADMediationBannerAdConfiguration *)adConfiguration
+                   completionHandler:(GADMediationBannerLoadCompletionHandler)completionHandler {
+  if ([GADMAdapterAppLovinUtils isChildUser]) {
+    completionHandler(nil, GADMAdapterAppLovinChildUserError());
+    return;
+  }
+
+  NSString *SDKKey =
+      [GADMAdapterAppLovinUtils retrieveSDKKeyFromCredentials:adConfiguration.credentials.settings];
+  if (!SDKKey) {
+    NSError *error = GADMAdapterAppLovinErrorWithCodeAndDescription(
+        GADMAdapterAppLovinErrorMissingSDKKey, @"AppLovin SDK Key is missing.");
+    completionHandler(nil, error);
+    return;
+  }
+  __weak GADMediationAdapterAppLovin *weakSelf = self;
+  // In the case of waterfall, initialize Applovin SDK before loading ad.
+  [GADMAdapterAppLovinInitializer
+      initializeWithSDKKey:SDKKey
+         completionHandler:^(void) {
+           GADMediationAdapterAppLovin *strongSelf = weakSelf;
+           if (!strongSelf) {
+             return;
+           }
+
+           strongSelf->_waterfallBannerRenderer = [[GADMWaterfallAppLovinBannerRenderer alloc]
+               initWithAdConfiguration:adConfiguration];
+           [strongSelf->_waterfallBannerRenderer loadAdWithCompletion:completionHandler];
+         }];
 }
 
 - (void)loadInterstitialForAdConfiguration: